Add jold_app to project
