+++ flowers-ol/experiment_manager_app/urls.py
@@ -16,14 +16,5 @@ urlpatterns = [
     path('start_task', views.start_task, name='start_task'),
     path('end_task', views.end_task, name='end_task'),
     path('thanks_page', views.thanks_page, name='thanks_page'),
-    path('end_session', views.end_session, name='end_session'),
-
-    # JOLD urls
-    path('JOLD/practice-LL', views.jold_start_ll_practice, name='jold_start_ll_practice'),
-    path('JOLD/save-trial', views.jold_save_ll_trial, name='jold_save_ll_trial'),
-    path('JOLD/close-LL-practice', views.jold_close_ll_practice, name='jold_close_ll_practice'),
-    path('JOLD/close-postsess-questionnaire', views.jold_close_postsess_questionnaire,
-         name='jold_close_postsess_questionnaire'),
-    path('JOLD/close-session', views.jold_free_choice, name='jold_free_choice'),
-    path('JOLD/close-session/<int:choice>/', views.jold_free_choice, name='jold_free_choice')
+    path('end_session', views.end_session, name='end_session')
 ]
