--- flowers-ol/experiment_manager_app/views/__init__.py
@@ -1,3 +1,0 @@
-from .interface import *
-from .common import *
-from .jold import *
