+++ flowers-ol/flowers-ol/urls.py
@@ -22,5 +22,6 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('experiment_manager_app.urls')),
     path('', include('mot_app.urls')),
-    path('', include('survey_app.urls'))
+    path('', include('survey_app.urls')),
+    path('', include('jold_app.urls'))
 ]
