+++ flowers-ol/jold_app/static/js/app_LL/utilities.js
